+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Sync replies in followers-only and subscribers-only conversations between participants.
 - Add `poll` attribute to Mastodon API `Status` entity.
 - Implemented `/api/v1/polls/{poll_id}/votes` API endpoint.
+- Added `--as-user` parameter to `import-object` command.
 
 ### Changed